Converted to DEM single precision because that is what is needed
(cherry picked from commit 9716eacf3bdc86626bc7a0ee553db90a8fcc6a0c)

+++ dggs/avalanche/domain.py
@@ -292,7 +292,7 @@ def dem_example():
 
     grid_info = gisutil.RasterInfo('', ni, nj, [0.0, 2.0, 0.0, 0.0, 0.0, 2.0])
 
-    return grid_info,dem,-5
+    return grid_info,dem.astype(np.single),-5
 
 def main():
